Support hardware decoding.

+++ src/VideoInfo.h
@@ -14,7 +14,7 @@ struct VideoInfo {
     int width = 0;
     int height = 0;
     int duration = 0;
-    int size = 0;
+    int64_t size = 0;
     int bit_rate = 0;
     int num_hard_links = 0;
     long inode = 0;
